plot results of esttraion environment
Report of some graphical results of an agent trained using PPo  for 10000 iteration  in the BeamAngleEnv environment in TestTrain.py, on some perturbation of a certain VOI(square blue points), target (black round points) configurations. We see how the agent is able to understand he position of the target and to irradiate without doing too much damage on VOI's.